requirements.txt
asgiref==3.4.1
astroid==2.11.7
dill==0.3.4
Django==3.2.15
django-cors-headers==3.10.1
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
pkg_resources==0.0.0
platformdirs==2.4.0
psycopg2-binary==2.9.3
pylint==2.13.9
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
wrapt==1.14.1